backend/ChatAPI/requirements.txt
pycryptodome
requests
fastapi
sqlalchemy
pymysql
uvicorn
debugpy
pydantic
pydantic[email]
httpx
cryptography
python-multipart
kafka-python
aiokafka
python-dotenv
asyncmy
uvicorn[standard]
python-jose
passlib
bcrypt
PyJWT